requirements.txt
fastapi
uvicorn
pycryptodome
requests
pytest
httpx
